bug fix for python<3.11
